refactor: extract predefTests into it's own suite
